// src/generated/api.ts
export interface paths {
    "/oauth/token": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * Get OAuth token (unauthenticated api)
         * @description Retrieve an OAuth Access token for specified a refresh token. Take `id_token` value from **response** as Oauth access token value for all subsequent authorization API calls.
         */
        post: operations["getoauthtoken"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/apis/ngpc.rxt.io/v1/regions": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get all regions
         * @description List objects of kind Region to provide all available regions.
         */
        get: operations["getregion"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/apis/ngpc.rxt.io/v1/regions/{name}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get region by name
         * @description Get the specified Region.
         */
        get: operations["readNgpcRxtIoV1Region"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/apis/ngpc.rxt.io/v1/serverclasses": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get all serverclasses
         * @description List objects of kind ServerClass to provide all available serverclasses.
         */
        get: operations["getserverclass"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/apis/ngpc.rxt.io/v1/serverclasses/{name}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get serverclass by name
         * @description Read the specified ServerClass
         */
        get: operations["readNgpcRxtIoV1ServerClass"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/apis/auth.ngpc.rxt.io/v1/organizations": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * List all organizations
         * @description List all organizations along with their details, including the associated namespaces.
         */
        get: operations["helper-getorganizationslist"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/apis/ngpc.rxt.io/v1/namespaces/{namespace}/cloudspaces": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get all cloudspaces
         * @description Retrieve all CloudSpaces from the specified namespace.
         */
        get: operations["listNgpcRxtIoV1NamespacedCloudSpace"];
        put?: never;
        /**
         * Create cloudspace
         * @description Create a CloudSpace in the specified namespace.
         */
        post: operations["createNgpcRxtIoV1NamespacedCloudSpace"];
        /**
         * Delete all cloudspaces
         * @description Delete CloudSpaces from the specified namespace.
         */
        delete: operations["deleteNgpcRxtIoV1CollectionNamespacedCloudSpace"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/apis/ngpc.rxt.io/v1/namespaces/{namespace}/cloudspaces/{name}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get cloudspace by name
         * @description Retrieves the details of a CloudSpace by its name from the specified namespace.
         */
        get: operations["readNgpcRxtIoV1NamespacedCloudSpace"];
        put?: never;
        post?: never;
        /**
         * Delete cloudspace by name
         * @description Delete a CloudSpace by its name from the specified namespace.
         */
        delete: operations["deleteNgpcRxtIoV1NamespacedCloudSpace"];
        options?: never;
        head?: never;
        /**
         * Edit cloudspace by name
         * @description Update a CloudSpace by its name from the specified namespace.
         */
        patch: operations["patchNgpcRxtIoV1NamespacedCloudSpace"];
        trace?: never;
    };
    "/apis/ngpc.rxt.io/v1/namespaces/{namespace}/spotnodepools": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get all spotnodepools
         * @description Retrieve all SpotNodePools from the specified namespace.
         */
        get: operations["listNgpcRxtIoV1NamespacedSpotNodePool"];
        put?: never;
        /**
         * Create spotnodepool
         * @description Create a SpotNodePool in the specified namespace.
         */
        post: operations["createNgpcRxtIoV1NamespacedSpotNodePool"];
        /**
         * Delete all spotnodepools
         * @description Delete all SpotNodePools from the specified namespace.
         */
        delete: operations["deleteNgpcRxtIoV1CollectionNamespacedSpotNodePool"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/apis/ngpc.rxt.io/v1/namespaces/{namespace}/spotnodepools/{name}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get spotnodepool by name
         * @description Retrieves the details of a SpotNodePool by its name from the specified namespace.
         */
        get: operations["readNgpcRxtIoV1NamespacedSpotNodePool"];
        put?: never;
        post?: never;
        /**
         * Delete spotnodepool by name
         * @description Delete a SpotNodePool by its name from the specified namespace.
         */
        delete: operations["deleteNgpcRxtIoV1NamespacedSpotNodePool"];
        options?: never;
        head?: never;
        /**
         * Edit spotnodepool by name
         * @description Update a SpotNodePool by its name from the specified namespace.
         */
        patch: operations["patchNgpcRxtIoV1NamespacedSpotNodePool"];
        trace?: never;
    };
    "/apis/ngpc.rxt.io/v1/namespaces/{namespace}/ondemandnodepools": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get all ondemandnodepools
         * @description Retrieve all OnDemandNodePools from the specified namespace.
         */
        get: operations["listNgpcRxtIoV1NamespacedOnDemandNodePool"];
        put?: never;
        /**
         * Create ondemandnodepool
         * @description Create an OnDemandNodePool in the specified namespace.
         */
        post: operations["createNgpcRxtIoV1NamespacedOnDemandNodePool"];
        /**
         * Delete all ondemandnodepools
         * @description Delete all OnDenamdNodePools from the specified namespace.
         */
        delete: operations["deleteNgpcRxtIoV1CollectionNamespacedOnDemandNodePool"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/apis/ngpc.rxt.io/v1/namespaces/{namespace}/ondemandnodepools/{name}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get ondemandnodepool by name
         * @description Retrieves the details of an OnDemandNodePool by its name from the specified namespace.
         */
        get: operations["readNgpcRxtIoV1NamespacedOnDemandNodePool"];
        put?: never;
        post?: never;
        /**
         * Delete ondemandnodepool by name
         * @description Delete an OnDemandNodePool by its name from the specified namespace.
         */
        delete: operations["deleteNgpcRxtIoV1NamespacedOnDemandNodePool"];
        options?: never;
        head?: never;
        /**
         * Edit ondemandnodepool by name
         * @description Update a OnDemandNodePool by its name from the specified namespace.
         */
        patch: operations["patchNgpcRxtIoV1NamespacedOnDemandNodePool"];
        trace?: never;
    };
// ...
}